Clean up repo for contributors: remove dead modules, add community files
- Remove deleted modules: ai_gguf, backend_manager, backend_plugin_api, app
- Add LICENSE (MIT), CONTRIBUTING.md, GitHub issue/PR templates
- Add .gitignore entries for Claude Code files (CLAUDE.md, .claude/, docs/plans/)
- Clean SECURITY.md: remove app module references, renumber findings
- Remove dead code: extractTarGzWithBuiltinGzip(), stale comments
- Delete boilerplate ExampleTest files across all modules
- Clarify commons-compress/xz dependency requirements for consumers
- Major ai_sd refactor: modularized C++ pipeline, image processing modules,
  runtime setup progress, adaptive buffers, safety checker, upscaler JNI
- gguf_lib: expanded JNI SDK with embeddings, RAG, tool calling, character engine

+++ ai_sd/src/main/cpp/src/utils/safetensor_reader.h
@@ -66,10 +66,10 @@ class SafeTensorReader {
     file.close();
   }
 
-  int calculate_tensor_size(const std::vector<int> &shape) {
-    int size = 1;
+  int64_t calculate_tensor_size(const std::vector<int> &shape) {
+    int64_t size = 1;
     for (int dim : shape) {
-      size *= dim;
+      size *= static_cast<int64_t>(dim);
     }
     return size;
   }
@@ -96,7 +96,7 @@ class SafeTensorReader {
       throw std::runtime_error("Unsupported tensor dtype: " + info.dtype);
     }
 
-    int tensor_size = calculate_tensor_size(info.shape);
+    int64_t tensor_size = calculate_tensor_size(info.shape);
     long data_start = info.data_offsets[0];
     long data_end = info.data_offsets[1];
